backend/services/email-service/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
alembic==1.14.0
anyio==4.6.2.post1
apispec==6.7.1
appnope==0.1.4
argcomplete==3.5.1
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
Authlib==1.3.2
babel==2.16.0
beautifulsoup4==4.10.0
blinker==1.9.0
bs4==0.0.2
cachelib==0.9.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==2.0.12
click==8.1.7
clickclick==20.10.2
colorama==0.4.6
colorlog==6.9.0
comm==0.2.2
ConfigUpdater==3.2
confluent-kafka==2.6.1
connexion==2.14.2
contourpy==1.3.1
cron-descriptor==1.4.5

cryptography==43.0.3
cycler==0.12.1
debugpy==1.8.8
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.9
dnspython==2.7.0
email_validator==2.2.0
executing==2.1.0
falcon==4.0.2
falcon-cors==1.1.7
# Flask==3.1.0
# Flask-AppBuilder==4.5.2
# Flask-Babel==2.0.0
# Flask-Caching==2.3.0
# Flask-JWT-Extended==4.6.0
# Flask-Limiter==3.8.0
# Flask-Login==0.6.3
# Flask-Session==0.5.0
# Flask-SQLAlchemy==2.5.1
# Flask-WTF==1.2.2
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.10.0
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.1
google-cloud-pubsub==2.27.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-re2==1.1.20240702
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
# grpc-google-iam-v1==0.13.1
# grpcio==1.68.0
# grpcio-status==1.68.0
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.0
idna==3.10
importlib-metadata==6.11.0
importlib_resources==6.4.5
inflection==0.5.1
ipykernel==6.29.5
ipython==8.29.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jmespath==0.10.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kafka-python==2.0.2
kafka-python-ng==2.2.3
kiwisolver==1.4.7
lazy-object-proxy==1.10.0
limits==3.13.0
linkify-it-py==2.0.3
lockfile==0.12.2
Mako==1.3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.28.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.5.0
multidict==6.1.0
nest-asyncio==1.6.0
numpy==2.1.3
oauth2client==4.1.3
oauthlib==3.2.2
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
ordered-set==4.1.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pendulum==3.0.0
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
prison==0.2.1
prompt_toolkit==3.0.48
propcache==0.2.0
proto-plus==1.25.0
# protobuf==5.28.3
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.2.0
python-daemon==3.1.0
python-dateutil==2.9.0.post0
python-dotenv==0.19.0
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2021.1
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.26.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.9.4
rich-argparse==1.6.0
rpds-py==0.20.1
rsa==4.9
scikit-learn==1.5.2
scipy==1.14.1
setproctitle==1.3.3
setuptools==75.6.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.1
stack-data==0.6.3
tabulate==0.9.0
tenacity==8.5.0
termcolor==2.5.0
text-unidecode==1.3
threadpoolctl==3.5.0
time-machine==2.16.0
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
uc-micro-py==1.0.3
universal_pathlib==0.2.5
uritemplate==4.1.1
urllib3==1.26.17
wcwidth==0.2.13
# Werkzeug==3.1.3
wirerope==0.4.7
wrapt==1.16.0
WTForms==3.2.1
yarl==1.17.1
zipp==3.20.2
